# common/swagger.yaml
openapi: 3.0.3
info:
  title: Discord Bot Builder API
  description: API for building and managing Discord bots
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Enter JWT token in format: Bearer <token>"
...
paths:
...
  /v1/bots/start:
    post:
      tags:
        - Bots
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                botId:
                  type: string
              required:
                - botId
              additionalProperties: false
        required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
  /v1/bots/stop:
    post:
      tags:
        - Bots
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                botId:
                  type: string
              required:
                - botId
              additionalProperties: false
        required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
  /v1/bots/{botId}:
    get:
      tags:
        - Bots
      parameters:
        - schema:
            type: string
          in: path
          name: botId
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
    put:
      tags:
        - Bots
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - not: {}
                - type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    token:
                      type: string
                    active:
                      type: boolean
                    ownerId:
                      type: string
                    prefix:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
                    description:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
                    icon:
                      type: string
                      nullable: true
                    serverId:
                      type: string
                  required:
                    - id
                    - name
                    - token
                    - active
                    - ownerId
                    - prefix
                    - createdAt
                    - updatedAt
                    - description
                    - tags
                    - icon
                    - serverId
                  additionalProperties: false
      parameters:
        - schema:
            type: string
          in: path
          name: botId
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - Bots
      parameters:
        - schema:
            type: string
          in: path
          name: botId
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
  /v1/bots/{botId}/sync:
    put:
      tags:
        - Bots
      parameters:
        - schema:
            type: string
          in: path
          name: botId
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
servers:
  - url: http://localhost:3000
    description: Development server